Created ClientSection, InternTestimonialsSection and CareersPageCarousel components for DRYer code
